apps/db/requirements.txt
curl-cffi
asyncpg
nba_api
pandas
pydantic-settings
tenacity
rich
